+++ lib/FastTransformer.php
@@ -1,8 +1,9 @@
 <?php declare(strict_types=1);
 namespace Robo\PhpMinify;
 
-use GuzzleHttp\{Client};
+use Symfony\Component\HttpClient\{HttpClient};
 use Symfony\Component\Process\{Process};
+use Symfony\Contracts\HttpClient\{HttpClientInterface};
 
 /** Removes comments and whitespace from a PHP script, by calling a Web service. */
 class FastTransformer implements Transformer {
@@ -12,6 +13,9 @@ class FastTransformer implements Transformer {
 
   /** @var string The path to the PHP executable. */
   private string $executable;
+
+  /** @var HttpClientInterface The HTTP client. */
+  private HttpClientInterface $http;
 
   /** @var int The port that the PHP process is listening on. */
   private int $port = -1;
@@ -26,6 +30,7 @@ class FastTransformer implements Transformer {
   function __construct(string $executable = 'php') {
     assert(mb_strlen($executable) > 0);
     $this->executable = $executable;
+    $this->http = HttpClient::create();
   }
 
   /** Closes this transformer and releases any resources associated with it. */
@@ -52,8 +57,9 @@ class FastTransformer implements Transformer {
    */
   function listen(): int {
     if (!$this->isListening()) {
+      $address = static::address;
       $this->port = $this->getPort();
-      $this->process = new Process([$this->executable, '-S', static::address.":{$this->port}", '-t', __DIR__]);
+      $this->process = new Process([$this->executable, '-S', "$address:{$this->port}", '-t', __DIR__]);
       $this->process->start();
       sleep(1);
     }
@@ -69,9 +75,9 @@ class FastTransformer implements Transformer {
   function transform(string $script): string {
     assert(mb_strlen($script) > 0);
     $address = static::address;
-    $file = rawurlencode((string) realpath($script));
     $port = $this->listen();
-    return (string) (new Client)->get("http://$address:$port/Server.php?file=$file")->getBody();
+    $query = ['file' => (string) realpath($script)];
+    return $this->http->request('GET', "http://$address:$port/Server.php", ['query' => $query])->getContent();
   }
 
   /**