+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version [2.1.0](https://github.com/cedx/robo-php-minify/compare/v2.0.0...v2.1.0)
+- Replaced the [Guzzle HTTP client](http://docs.guzzlephp.org) by the [Symfony one](https://symfony.com/doc/current/components/http_client.html) in `FastTransformer` class.
 - Updated the package dependencies.
 
 ## Version [2.0.0](https://github.com/cedx/robo-php-minify/compare/v1.0.0...v2.0.0)